requirements.txt
asgiref==3.5.0
bleach==5.0.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.2.13
django-ckeditor==6.3.2
django-environ==0.8.1
django-heroku==0.3.1
django-jazzmin==2.5.0
django-js-asset==2.0.0
docopt==0.6.2
gunicorn==20.1.0
idna==3.3
mysqlclient==2.1.0
Pillow==9.0.1
pipreqs==0.4.11
psycopg2==2.9.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.9
webencodings==0.5.1
whitenoise==6.1.0
yarg==0.1.9
